moved BlockTracker to synchronization package
